Prepare for CRAN resubmission.
Original comments:
------------------

License components with restrictions and base license permitting such:
    GPL-2 | file LICENSE

We do not need "| file LICENSE" and the file as these are part of R. This is only needed in case of attribution requirements or other possible restrictions.
Hence please omit it.

Please add \value to .Rd files regarding exported methods and explain the functions results in the documentation. Please write about the structure of the output (class) and also what the output means. (If a function does not return a value, please document that too, e.g. \value{No return value, called for side effects} or similar)
Missing Rd-tags:
     geom_glowpath.Rd: \value
     geom_glowpoint.Rd: \value
     geom_shadowpath.Rd: \value
     geom_shadowpoint.Rd: \value
     ggshadow-params.Rd: \value
     scale_colour_hue.Rd: \value
     scale_colour_steps.Rd: \value
     scale_continuous.Rd: \value

If possible: Please add some more small executable examples in your Rd-files to illustrate the use of the exported function but also enable automatic testing.